apps/api/pyproject.toml
[project]
name = "mao-api"
version = "0.1.0"
description = "Multi-Agent Orchestration Platform — FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    # ── Web framework ──────────────────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.1.0",

    # ── LangGraph / LangChain ─────────────────────────────────────────
    "langgraph>=0.2.60",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain>=0.3.10",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.10",
    "langmem>=0.0.10",

    # ── Anthropic ─────────────────────────────────────────────────────
    "anthropic>=0.42.0",

    # ── Memory / Knowledge Graph ──────────────────────────────────────
    "mem0ai[graph]>=0.1.50",
    "kuzu>=0.7.0",

    # ── AG-UI ────────────────────────────────────────────────────────
    "ag-ui-langgraph>=0.1.0",

    # ── Observability ─────────────────────────────────────────────────
    "langfuse>=2.0.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-httpx>=0.49b0",

    # ── Database / Cache ──────────────────────────────────────────────
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "redis[hiredis]>=5.2.0",

    # ── MCP ───────────────────────────────────────────────────────────
    "mcp>=1.2.0",
    "langchain-mcp-adapters>=0.1.0",

    # ── Config / Validation ───────────────────────────────────────────
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",

    # ── HTTP / Async ──────────────────────────────────────────────────
    "httpx>=0.28.0",
    "aiofiles>=24.1.0",

    # ── Utilities ─────────────────────────────────────────────────────
    "orjson>=3.10.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-aiofiles>=24.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
addopts = "--cov=src --cov-report=term-missing -v"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-aiofiles>=24.1.0",
]
